requirements.txt
streamlit>=1.43.0
requests>=2.32.3
pandas>=2.2.3
python-qbittorrent>=0.4.3